TubeGuard_GFC_System/requirements.txt
# requirements.txt
# TubeGuard-AI 依赖列表

# 核心图像处理
opencv-python>=4.8.0      # OpenCV: 用于图像读取、基础几何计算、Canny边缘检测等
numpy>=1.24.0             # NumPy: 矩阵运算，亚像素算法核心依赖
scikit-image>=0.21.0      # Scikit-image: 提供一些高级图像处理算法

# 深度学习框架
torch>=2.0.0              # PyTorch: SegFormer, PatchCore, YOLO 的基础框架
torchvision>=0.15.0       # TorchVision: 常用视觉模型和变换

# 模型架构与工具
ultralytics>=8.1.0        # YOLOv8/v12: 用于 Stage 3 (Stage B) 的目标检测
timm>=0.9.0               # PyTorch Image Models: 用于加载 SegFormer backbone
anomalib>=0.7.0           # Anomalib: 提供 PatchCore 等无监督异常检测算法实现 (可选)
faiss-cpu>=1.7.4          # FAISS: PatchCore Memory Bank 的高效检索 (必须)

# 工具库
PyYAML>=6.0               # PyYAML: 读取 config.yaml 配置文件
tqdm>=4.65.0              # TQDM: 进度条显示
matplotlib>=3.7.0         # Matplotlib: 用于绘制 debug_viz 图表
